poset tests are in separate files
