+++ src/posposet/poset_clotho_test.go
@@ -1,0 +1,11 @@
+package posposet
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPosetSimpleClotho(t *testing.T) {
+	_ = assert.New(t)
+}
